pyproject.toml
[tool.poetry]
name = "swapi-a-star-wars-api"
version = "0.1.0"
description = ""
authors = ["Shauryadeep Chaudhuri <shauryadeepc@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "swapi_a_star_wars_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
petl = "^1.7.12"
requests = "^2.28.2"
djangorestframework = "^3.14.0"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
